move schema customization to createSchemaCustomization
