+++ packages/gatsby-source-wordpress/src/utils/gql.js
@@ -1,0 +1,3 @@
+const gql = ([string]) => string
+
+export default gql
